+++ src/config/light.rs
@@ -2,8 +2,8 @@
 #[serde(tag = "type")]
 pub enum Light {
     PointLight {
-        origin: [f64; 3],
-        color: [f64; 3],
-        intensity: f64,
+        origin: [f32; 3],
+        color: [f32; 3],
+        intensity: f32,
     },
 }
