+++ src/lights/point_light.rs
@@ -4,11 +4,11 @@ use color::Color;
 pub struct PointLight {
     pub origin: Point3,
     pub color: Color,
-    intensity: f64,
+    intensity: f32,
 }
 
 impl PointLight {
-    pub fn new(origin: Point3, color: Color, intensity: f64) -> PointLight {
+    pub fn new(origin: Point3, color: Color, intensity: f32) -> PointLight {
         PointLight {
             origin: origin,
             color: color,
@@ -16,7 +16,7 @@ impl PointLight {
         }
     }
 
-    pub fn intensity(&self, distance_to_light: f64) -> f64 {
+    pub fn intensity(&self, distance_to_light: f32) -> f32 {
         1.0 / (distance_to_light * distance_to_light) * self.intensity
     }
 }
